requirements.txt
autopep8==1.4.4
Babel==2.8.0
beautifulsoup4==4.9.1
boto3==1.10.49
botocore==1.13.50
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==3.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.2.0
django-fullurl==1.1
django-heroku==0.3.1
django-import-export==2.2.0
django-indian-numbers @ git+https://github.com/narenchoudhary/django-indian-numbers@fea14d68d03de7d27d2ad8b4abf5794d55b9b88a
django-model-utils==3.2.0
django-mssql==1.8
django-payu==0.5
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-storages==1.8
django-stubs==1.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==1.1.1
mypy==0.761
mypy-extensions==0.4.3
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.3
pep8==1.7.1
Pillow==6.0.0
psycopg2==2.8.5
pycodestyle==2.5.0
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
stripe==2.27.0
tablib==2.0.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.24.2
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
